Fix leak by ensuring both ends of target connection are closed when the client sends EOF.
As further explanation: the lduplex connection is used in local.dialPipe and
provided to service.tcpService as the connection to the "target" (in reality,
the consumer of the listener provided by the lantern package). The lduplex type
fakes the ability to independently close the write or read side of the
connection. In handling each connection, a pair of goroutines execute a
bidirectional copy from client to upstream and vice versa. If closing the write
side is a no-op, then the upstream-reader is never informed when the client has
sent EOF. This can result in leaked goroutines and file descriptors.

This was originally fixed in
https://github.com/getlantern/lantern-shadowsocks/pull/6

However, that fix resulted in a data race regarding the metrics.ProxyMetrics
instance associated with each handled connection. The bidirectional-copy
goroutines could still be running and writing to the metrics when the metrics
are read and passed to the caller-provided metrics.ShadowsocksMetrics instance.

This new fix ensures the proxy is informed when the client sends EOF and avoids
data races.